+++ NEWS.md
@@ -3,7 +3,8 @@
 * Initial release of `teal.slice` a package providing a filter module for `teal` applications.
 
 ## Changes (from behavior when functionality was part of `teal`)
-* The filter panel now displays a helpful message when data has no rows or columns. This change is from implementations made when `teal.slice` was still part of `teal`.
+* The filter panel now displays a helpful message when data has no rows or columns.
+* `FilteredData` now stores whether its datasets had a reproducibility check or not.
 
 ### Breaking changes
 * `default_filter` has been deprecated - use `list()` instead for a default filter.